Move extensions.py to a package, and rename to pydantic.py
